refactor: refactor class MeshUtils to namespace

+++ src/MeshUtils.cpp
@@ -121,6 +121,6 @@ Equation::Plane MeshUtils::computeFacePlane(const Mesh& mesh, const Mesh::FaceHa
 	Mesh::Normal normal{ OpenMesh::cross(vec_1, vec_2) };
 	Mesh::Point p{ mesh.point(mesh.to_vertex_handle(heh)) }; //a, b and c
 
-	int d{ normal[0] * p[0] + normal[1] * p[1] + normal[2] * p[2] };
+	float d{ normal[0] * p[0] + normal[1] * p[1] + normal[2] * p[2] };
 	return Equation::Plane{ normal[0], normal[1], normal[2], d };
 }
